+++ scilab/CHANGES.md
@@ -359,6 +359,7 @@ Bug Fixes
 * [#16841](https://gitlab.com/scilab/scilab/-/issues/16841): Zoom on multiaxes could desynchronize axes.
 * [#16849](https://gitlab.com/scilab/scilab/-/issues/16849): toJSON("\\") must return "\\\\" instead of "\\\".
 * [#16865](https://gitlab.com/scilab/scilab/-/issues/16865): Vector extraction with single hypermatrix index yielded a vector.
+* [#16593](https://gitlab.com/scilab/scilab/-/issues/16593): After L = list(1,2); , for o = L(3:$), o, end crashes Scilab.
 * [#16029](https://bugzilla.scilab.org/16029): tick marks were extending out of range for small scale plots 
 * [#16653](https://bugzilla.scilab.org/16653): fromJSON() crashed with a non existing file.
 * [#16713](https://bugzilla.scilab.org/16713): Automatic stretching of Axes box in 3d view was not configurable.